Force Zimbra cron jobs to silently exit without doing anything if run on a cluster standby host.  The scripts know it's running on a standby host when /opt/zimbra/conf/localconfig.xml file is not found. (because the file is on SAN and standby host doesn't have SAN mounted)

+++ ZimbraServer/src/bin/zmroll_catalina.sh
@@ -24,6 +24,10 @@
 # ***** END LICENSE BLOCK *****
 # 
 
+if [ ! -f /opt/zimbra/conf/localconfig.xml ]; then
+    exit 0
+fi
+
 DATE=`date +%Y%m%d%H%M`
 cp -f /opt/zimbra/tomcat/logs/catalina.out \
 	/opt/zimbra/tomcat/logs/catalina.out.$DATE
